Move EfUserContextAccessor namespace, updated service collection extension name.
